NYT_SCRAPING_WoolworthsShop/NTY_PIPELINE/PART2/requirements.txt
apache-airflow-providers-google
pyarrow
pandas
request